+++ packages/std/src/math/int256.rs
@@ -255,7 +255,7 @@ impl Int256 {
     }
 
     #[must_use = "this returns the result of the operation, without modifying the original"]
-    pub fn abs_diff(self, other: Self) -> Uint256 {
+    pub const fn abs_diff(self, other: Self) -> Uint256 {
         Uint256(self.0.abs_diff(other.0))
     }
 }