+++ packages/std/src/math/uint512.rs
@@ -295,7 +295,7 @@ impl Uint512 {
     }
 
     #[must_use = "this returns the result of the operation, without modifying the original"]
-    pub fn abs_diff(self, other: Self) -> Self {
+    pub const fn abs_diff(self, other: Self) -> Self {
         Self(self.0.abs_diff(other.0))
     }
 }