+++ CHANGELOG.md
@@ -13,7 +13,9 @@ and this project adheres to
 - cosmwasm-std: Make `abs_diff` const for `Uint{256,512}` and
   `Int{64,128,256,512}`. It is now const for all integer types.
 - cosmwasm-std: Implement `TryFrom<Decimal256>` for `Decimal` ([#1832])
+- cosmwasm-std: Add `StdAck`. ([#1512])
 
+[#1512]: https://github.com/CosmWasm/cosmwasm/issues/1512
 [#1799]: https://github.com/CosmWasm/cosmwasm/pull/1799
 [#1806]: https://github.com/CosmWasm/cosmwasm/pull/1806
 [#1832]: https://github.com/CosmWasm/cosmwasm/pull/1832