+++ CHANGELOG.md
@@ -13,10 +13,14 @@ and this project adheres to
 - cosmwasm-std: Make `abs_diff` const for `Uint{256,512}` and
   `Int{64,128,256,512}`. It is now const for all integer types.
 - cosmwasm-std: Implement `TryFrom<Decimal256>` for `Decimal` ([#1832])
+- cosmwasm-std: Add new imports `db_next_{key, value}` for iterating storage
+  keys / values only and make `Storage::{range_key, range_value}` more
+  efficient. This requires the `cosmwasm_1_4` feature to be enabled. ([#1834])
 
 [#1799]: https://github.com/CosmWasm/cosmwasm/pull/1799
 [#1806]: https://github.com/CosmWasm/cosmwasm/pull/1806
 [#1832]: https://github.com/CosmWasm/cosmwasm/pull/1832
+[#1834]: https://github.com/CosmWasm/cosmwasm/pull/1834
 
 ### Changed