+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.4.0-beta.1] - 2023-08-29
+
 ### Added
 
 - cosmwasm-std: Implement `Not` for `Uint{64,128,256}` ([#1799]).
@@ -1851,7 +1853,9 @@ Some main points:
 
 All future Changelog entries will reference this base
 
-[unreleased]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.3...HEAD
+[unreleased]: https://github.com/CosmWasm/cosmwasm/compare/v1.4.0-beta.1...HEAD
+[1.4.0-beta.1]:
+  https://github.com/CosmWasm/cosmwasm/compare/v1.3.3...v1.4.0-beta.1
 [1.3.3]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.2...v1.3.3
 [1.3.2]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.1...v1.3.2
 [1.3.1]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.0...v1.3.1